requirements.txt
pytorch-lightning<2.0.0
torchvision
torchviz
torchsummary
tensorboard
matplotlib
scikit-learn
imgaug
ipyplot
jupyter
psutil
pandas
seaborn
umap-learn
numpy
bokeh



